[tkw] Remove wave ops and types
